whisper_stt/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
huggingface_hub
python-dotenv
